Extracted BookStoreControllerBase to enable sharing of some controller functionality
